Serve custom 404 page for any request
We want to direct users to the tutorials and examples on the Prototype
Kit website, and inform them that the previous location is no longer
valid. The easiest way to do this is to serve a custom 404 page that
shows for any request, pointing them towards the current documentation
site. This won't redirect automatically, but will give users enough
information to update their bookmarks or guidance as necessary.

Co-authored-by: domoscargin <brett.kyle@digital.cabinet-office.gov.uk>
Co-authored-by: NoraGDS <57447099+NoraGDS@users.noreply.github.com>